rad-clib: flag for unix builds
We flag a unix module for the ssh agent signer code. The same functions
are under non-unix, however they are defined as unimplemented. These
functions are re-exported at the top-level to satisfy compilation
builds.
