+++ rad-clib/Cargo.toml
@@ -32,4 +32,4 @@ features = ["std"]
 [dependencies.thrussh-agent]
 git = "https://github.com/FintanH/thrussh"
 branch = "generic-agent"
-default-features = false
+features = [ "tokio-agent" ]
