+++ rad-clib/src/keys/ssh.rs
@@ -3,28 +3,9 @@
 // This file is part of radicle-link, distributed under the GPLv3 with Radicle
 // Linking Exception. For full terms see the included LICENSE file.
 
-use std::{fmt, sync::Arc};
-
-use serde::{de::DeserializeOwned, Serialize};
 use thiserror::Error;
-use thrussh_agent::Constraint;
-use tokio::net::UnixStream;
 
-use librad::{
-    crypto::{
-        keystore::{
-            crypto::Crypto,
-            sign::ssh::{self, SshAgent},
-            Keystore as _,
-        },
-        BoxedSigner,
-        SomeSigner,
-    },
-    git::storage::{read, ReadOnly},
-    profile::Profile,
-};
-
-use crate::runtime;
+use librad::{crypto::keystore::sign::ssh, git::storage::read};
 
 #[derive(Debug, Error)]
 pub enum Error {
@@ -38,28 +19,12 @@ pub enum Error {
     StorageInit(#[from] read::error::Init),
 }
 
-pub fn signer(profile: &Profile) -> Result<BoxedSigner, Error> {
-    let storage = ReadOnly::open(profile.paths())?;
-    let peer_id = storage.peer_id();
-    tracing::trace!(peer=%peer_id, "obtaining signer for peer");
-    let agent = SshAgent::new((*peer_id.as_public_key()).into());
-    let signer = runtime::block_on(agent.connect::<UnixStream>())?;
-    Ok(SomeSigner {
-        signer: Arc::new(signer),
-    }
-    .into())
-}
+#[cfg(unix)]
+mod unix;
+#[cfg(unix)]
+pub use unix::*;
 
-pub fn add_signer<C>(profile: &Profile, crypto: C, constraints: &[Constraint]) -> Result<(), Error>
-where
-    C: Crypto,
-    C::Error: fmt::Debug + fmt::Display + Send + Sync + 'static,
-    C::SecretBox: Serialize + DeserializeOwned,
-{
-    let store = super::file_storage(profile, crypto);
-    let key = store.get_key().map_err(|err| Error::GetKey(err.into()))?;
-    Ok(runtime::block_on(ssh::add_key::<UnixStream>(
-        key.secret_key.into(),
-        constraints,
-    ))?)
-}
+#[cfg(not(unix))]
+mod win;
+#[cfg(not(unix))]
+pub use win::*;
