+++ test/src/test/integration.rs
@@ -8,3 +8,5 @@ mod git_helpers;
 mod librad;
 mod link_git_protocol;
 mod node_lib;
+#[cfg(unix)]
+mod rad_cli;
