+++ test/src/test/unit/rad_clib.rs
@@ -3,4 +3,5 @@
 // This file is part of radicle-link, distributed under the GPLv3 with Radicle
 // Linking Exception. For full terms see the included LICENSE file.
 
+#[cfg(unix)]
 mod keys;
